Move InstanceSupplier to core.service package
